# perf: keyword lookups in the tokenizer