--- a/go/vt/sqlparser/keywords_test.go
+++ b/go/vt/sqlparser/keywords_test.go
@@ -1,15 +1,15 @@
 package sqlparser
 
-import "testing"
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
 
 func TestKeywordTable(t *testing.T) {
 	for _, kw := range keywords {
 		lookup, ok := keywordLookupTable.LookupString(kw.name)
-		if !ok {
-			t.Fatalf("keyword %q failed to match", kw.name)
-		}
-		if lookup != kw.id {
-			t.Fatalf("keyword %q matched to %d (expected %d)", kw.name, lookup, kw.id)
-		}
+		require.Truef(t, ok, "keyword %q failed to match", kw.name)
+		require.Equalf(t, lookup, kw.id, "keyword %q matched to %d (expected %d)", kw.name, lookup, kw.id)
 	}
 }