… go/vt/sqlparser/testdata/django_queries.txt
… go/vt/sqlparser/token.go