@vmg 2021-03-05
@harshit-gangal: I've added `testify` to all the new test cases, but I haven't updated the benchmarks to use `testify` because the `require.NoError` check actually adds measurable overhead and screws with the measures (:sweat:).

Ready to merge. :ok_hand: