@vmg 2021-03-08
@deepthi I haven't tested it, I just noticed the `testify` issue when re-running the benchmarks. I don't think it's particularly worrisome, Testify still works great for normal testing, so we can fall back to a simple `if err != nil` in the very few places where we have tight loops in benchmarks.